# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Collaborative Markdown Editor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Document {
  id: string;
  title: string;
  content: string;
  tags: string[];
  shareToken?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDocumentRequest {
  title: string;
  content?: string;
  tags?: string[];
}

export interface UpdateDocumentRequest {
  title?: string;
  content?: string;
  tags?: string[];
}

export interface UpdateTagsRequest {
  tags: string[];
}

export interface ShareTokenResponse {
  shareToken: string;
  shareUrl: string;
}

export interface DocumentVersion {
  id: string;
  documentId: string;
  title: string;
  content: string;
  label?: string | null;
  createdAt: string;
}

export interface SaveVersionRequest {
  label?: string;
}

export interface SaveToGithubRequest {
  filename: string;
  content: string;
  description?: string;
  isPublic?: boolean;
  gistId?: string | null;
}

export interface SaveToGithubResponse {
  gistId: string;
  url: string;
  htmlUrl: string;
  isUpdate: boolean;
}

export interface ErrorResponse {
  error: string;
  message: string;
}