# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Collaborative Markdown Editor API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Document operations
  - name: versions
    description: Document version history
  - name: github
    description: GitHub integration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a new document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/share/{token}:
    get:
      operationId: getDocumentByShareToken
      tags: [documents]
      summary: Get a document by its share token (read-only)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentRequest"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/duplicate:
    post:
      operationId: duplicateDocument
      tags: [documents]
      summary: Duplicate a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Duplicated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/tags:
    patch:
      operationId: updateDocumentTags
      tags: [documents]
      summary: Update document tags
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTagsRequest"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/share:
    post:
      operationId: generateShareToken
      tags: [documents]
      summary: Generate a share token for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Share token generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareTokenResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: revokeShareToken
      tags: [documents]
      summary: Revoke share token for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/versions:
    get:
      operationId: listDocumentVersions
      tags: [versions]
      summary: List all saved versions of a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentVersion"
    post:
      operationId: saveDocumentVersion
      tags: [versions]
      summary: Save the current state as a version snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveVersionRequest"
      responses:
        "201":
          description: Created version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentVersion"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/versions/{versionId}:
    delete:
      operationId: deleteDocumentVersion
      tags: [versions]
      summary: Delete a document version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/versions/{versionId}/restore:
    post:
      operationId: restoreDocumentVersion
      tags: [versions]
      summary: Restore a document to a previous version
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restored document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /github/save:
    post:
      operationId: saveToGithub
      tags: [github]
      summary: Save document to GitHub Gist (create or update)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveToGithubRequest"
      responses:
        "200":
          description: Saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveToGithubResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Document:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        shareToken:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - tags
        - createdAt
        - updatedAt
    CreateDocumentRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - title
    UpdateDocumentRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
    UpdateTagsRequest:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
      required:
        - tags
    ShareTokenResponse:
      type: object
      properties:
        shareToken:
          type: string
        shareUrl:
          type: string
      required:
        - shareToken
        - shareUrl
    DocumentVersion:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        title:
          type: string
        content:
          type: string
        label:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - documentId
        - title
        - content
        - createdAt
    SaveVersionRequest:
      type: object
      properties:
        label:
          type: string
    SaveToGithubRequest:
      type: object
      properties:
        filename:
          type: string
        content:
          type: string
        description:
          type: string
        isPublic:
          type: boolean
        gistId:
          type: string
          nullable: true
      required:
        - filename
        - content
    SaveToGithubResponse:
      type: object
      properties:
        gistId:
          type: string
        url:
          type: string
        htmlUrl:
          type: string
        isUpdate:
          type: boolean
      required:
        - gistId
        - url
        - htmlUrl
        - isUpdate
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message