# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Collaborative Markdown Editor API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all documents
 */
export const ListDocumentsResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  shareToken: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListDocumentsResponse = zod.array(ListDocumentsResponseItem);

/**
 * @summary Create a new document
 */
export const CreateDocumentBody = zod.object({
  title: zod.string(),
  content: zod.string().optional(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Get a document by its share token (read-only)
 */
export const GetDocumentByShareTokenParams = zod.object({
  token: zod.coerce.string(),
});

export const GetDocumentByShareTokenResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  shareToken: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Get a document by ID
 */
export const GetDocumentParams = zod.object({
  id: zod.coerce.string(),
});

export const GetDocumentResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  shareToken: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a document
 */
export const UpdateDocumentParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateDocumentBody = zod.object({
  title: zod.string().optional(),
  content: zod.string().optional(),
  tags: zod.array(zod.string()).optional(),
});

export const UpdateDocumentResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  shareToken: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a document
 */
export const DeleteDocumentParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Duplicate a document
 */
export const DuplicateDocumentParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Update document tags
 */
export const UpdateDocumentTagsParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateDocumentTagsBody = zod.object({
  tags: zod.array(zod.string()),
});

export const UpdateDocumentTagsResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  shareToken: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Generate a share token for a document
 */
export const GenerateShareTokenParams = zod.object({
  id: zod.coerce.string(),
});

export const GenerateShareTokenResponse = zod.object({
  shareToken: zod.string(),
  shareUrl: zod.string(),
});

/**
 * @summary Revoke share token for a document
 */
export const RevokeShareTokenParams = zod.object({
  id: zod.coerce.string(),
});

export const RevokeShareTokenResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  shareToken: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary List all saved versions of a document
 */
export const ListDocumentVersionsParams = zod.object({
  id: zod.coerce.string(),
});

export const ListDocumentVersionsResponseItem = zod.object({
  id: zod.string(),
  documentId: zod.string(),
  title: zod.string(),
  content: zod.string(),
  label: zod.string().nullish(),
  createdAt: zod.date(),
});
export const ListDocumentVersionsResponse = zod.array(
  ListDocumentVersionsResponseItem,
);

/**
 * @summary Save the current state as a version snapshot
 */
export const SaveDocumentVersionParams = zod.object({
  id: zod.coerce.string(),
});

export const SaveDocumentVersionBody = zod.object({
  label: zod.string().optional(),
});

/**
 * @summary Delete a document version
 */
export const DeleteDocumentVersionParams = zod.object({
  id: zod.coerce.string(),
  versionId: zod.coerce.string(),
});

/**
 * @summary Restore a document to a previous version
 */
export const RestoreDocumentVersionParams = zod.object({
  id: zod.coerce.string(),
  versionId: zod.coerce.string(),
});

export const RestoreDocumentVersionResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  content: zod.string(),
  tags: zod.array(zod.string()),
  shareToken: zod.string().nullish(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Save document to GitHub Gist (create or update)
 */
export const SaveToGithubBody = zod.object({
  filename: zod.string(),
  content: zod.string(),
  description: zod.string().optional(),
  isPublic: zod.boolean().optional(),
  gistId: zod.string().nullish(),
});

export const SaveToGithubResponse = zod.object({
  gistId: zod.string(),
  url: zod.string(),
  htmlUrl: zod.string(),
  isUpdate: zod.boolean(),
});