# lib/api-zod/src/generated/types/documentVersion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Collaborative Markdown Editor API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentVersion {
  id: string;
  documentId: string;
  title: string;
  content: string;
  label?: string | null;
  createdAt: Date;
}