# lib/api-zod/src/generated/types/saveToGithubRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Collaborative Markdown Editor API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveToGithubRequest {
  filename: string;
  content: string;
  description?: string;
  isPublic?: boolean;
  gistId?: string | null;
}