# lib/api-zod/src/generated/types/updateDocumentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Collaborative Markdown Editor API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateDocumentRequest {
  title?: string;
  content?: string;
  tags?: string[];
}